app/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone-client
tiktoken
langchain
langchain-community
uvicorn
fastapi
langchain-openai

